Add pencil beam
